Remove several compiler warnings.

+++ src/main.c
@@ -396,7 +396,7 @@ static int setupLookaside(sqlite3 *db, void *pBuf, int sz, int cnt){
   /* The size of a lookaside slot needs to be larger than a pointer
   ** to be useful.
   */
-  if( sz<=sizeof(LookasideSlot*) ) sz = 0;
+  if( sz<=(int)sizeof(LookasideSlot*) ) sz = 0;
   if( cnt<0 ) cnt = 0;
   if( sz==0 || cnt==0 ){
     sz = 0;
